requirements.txt
alembic==1.12.0
appdirs==1.4.4
astroid==3.0.1
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
commonmark==0.9.1
config==0.5.1
cryptography==41.0.5
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
dockerpty==0.4.1
docopt==0.6.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-CLI==0.4.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
flask-wtf==1.2.1
git-filter-repo==2.38.0
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
paramiko==3.3.1
platformdirs==3.11.0
pluggy==1.3.0
pycparser==2.21
pygitguardian==1.11.0
Pygments==2.16.1
PyJWT==2.6.0
pylint==3.0.2
pymssql==2.2.9
PyNaCl==1.5.0
pyodbc==5.0.1
pyrsistent==0.20.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
rich==12.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlalchemy-orm==1.2.10
texttable==1.7.0
tomli==2.0.1
tomlkit==0.12.2
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.0.7
websocket-client==0.59.0
werkzeug==3.0.1
wtforms==3.1.0
zipp==3.17.0
